+++ tests/integration/base.py
@@ -1,3 +1,6 @@
+from pyspark.sql.types import StringType
+
+from sparkle.utils import absolute_path
 from sparkle import SparkleContext
 
 
@@ -8,3 +11,13 @@ class _TestContext(SparkleContext):
                 'com.databricks:spark-csv_2.10:1.4.0',
                 'org.apache.spark:spark-streaming-kafka_2.10:1.6.1',
                 ]
+
+    jars = [
+        absolute_path(__file__, 'resources', 'brickhouse-0.7.1.jar'),
+        absolute_path(__file__, 'resources', 'mysql-connector-java-5.1.39-bin.jar'),
+    ]
+
+    udfs = {
+        'collect_max': 'brickhouse.udf.collect.CollectMaxUDAF',
+        'length_of_text': (lambda text: len(text), StringType())
+    }